src/app/questions/customer/core/address/address.guard.ts
import { Injectable } from '@angular/core';
import {
  CanActivate,
  ActivatedRouteSnapshot,
  RouterStateSnapshot
} from '@angular/router';
import { Observable } from 'rxjs/Observable';
import { CustomerRoutingService } from '../customer-routing.service';

@Injectable()
export class CanActivateQuestionsCustomerAddress implements CanActivate {
  public canActivate(next: ActivatedRouteSnapshot,
                     state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
    return this.routingService.validAddress();
  }

  constructor(private routingService: CustomerRoutingService) { }
}

